Se agregan validaciones de roles de usuario

+++ src/main/java/com/duoc/services/ComentarioServiceImpl.java
@@ -1,7 +1,9 @@
 package com.duoc.services;
 
+import com.duoc.adapter.controller.UsuarioClient;
 import com.duoc.dto.ComentarioDTO;
-import com.duoc.dto.PublicacionDTO;
+import com.duoc.dto.UsuarioDTO;
+import com.duoc.enums.UserRole;
 import com.duoc.exceptions.ComentarioNotFoundException;
 import com.duoc.exceptions.IllegalNumberException;
 import com.duoc.exceptions.UsuarioNotFoundException;
@@ -9,24 +11,26 @@ import com.duoc.mapper.ComentarioMapper;
 import com.duoc.model.ComentarioEntity;
 import com.duoc.repositories.ComentarioRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class ComentarioServiceImpl implements ComentarioService{
 
     private final ComentarioRepository comentarioRepository;
     private final ComentarioMapper comentarioMapper;
+    private final UsuarioClient usuarioClient;
 
     @Override
     public List<ComentarioDTO> getComentariosByIdPublicacion(Long idPublicacion) {
-        List<ComentarioEntity> comentariosPub = comentarioRepository.findByIdPublicacion(idPublicacion);
-
-        if(comentariosPub.isEmpty()) {
-            throw new ComentarioNotFoundException("No hay comentarios en la publicación");
-        }
+        List<ComentarioEntity> comentariosPub = comentarioRepository.findAllByIdPublicacion(idPublicacion)
+                .orElseThrow(() -> new ComentarioNotFoundException("No hay comentarios en la publicación"));
 
         return comentariosPub.stream()
                 .map(comentarioMapper::comentarioEntityToDto)
@@ -49,37 +53,91 @@ public class ComentarioServiceImpl implements ComentarioService{
     @Override
     public ComentarioDTO modificarComentario(ComentarioDTO comentarioDTO) {
 
-        if (comentarioDTO.getIdPublicacion() <= 0 && comentarioDTO.getIdUsuario() <= 0) {
+        if (comentarioDTO.getIdPublicacion() <= 0 || comentarioDTO.getIdUsuario() <= 0) {
             throw new IllegalNumberException("El ID debe ser positivo y no nulo");
         }
 
-        return comentarioRepository.findById(comentarioDTO.getIdComentario())
-                .map(comentarioEntity -> {
-                    comentarioEntity.setContenido(comentarioDTO.getContenido());
+        UsuarioDTO usuarioDTO = Optional.ofNullable(usuarioClient.obtenerUsuario(comentarioDTO.getIdUsuario()).getBody())
+                .orElseThrow(() -> new UsuarioNotFoundException("El usuario con ID " + comentarioDTO.getIdUsuario() + " no existe."));
 
-                    ComentarioEntity comentarioEntityModificado = comentarioRepository.save(comentarioEntity);
-                    return comentarioMapper.comentarioEntityToDto(comentarioEntityModificado);
-                })
-                .orElseThrow(() -> new ComentarioNotFoundException("Comentario no encontrado con ID: " + comentarioDTO.getContenido()));
+        if (!comentarioRepository.existsById(comentarioDTO.getIdComentario())) {
+            throw new ComentarioNotFoundException("El comentario con ID " + comentarioDTO.getIdComentario() + " no existe.");
+        }
+
+        ComentarioEntity comentario;
+
+        if (UserRole.ADMIN.equals(usuarioDTO.getRole()) || UserRole.MODERATOR.equals(usuarioDTO.getRole())) {
+            comentario = comentarioRepository.findById(comentarioDTO.getIdComentario())
+                    .orElseThrow(() -> new ComentarioNotFoundException("Comentario no encontrado con ID: " + comentarioDTO.getIdComentario()));
+        } else {
+            comentario = comentarioRepository.findByIdComentarioAndIdUsuario(comentarioDTO.getIdComentario(), comentarioDTO.getIdUsuario())
+                    .orElseThrow(() -> new ComentarioNotFoundException("Comentario no encontrado con ID: " + comentarioDTO.getIdComentario()));
+        }
+
+        comentario.setContenido(comentarioDTO.getContenido());
+
+        return comentarioMapper.comentarioEntityToDto(comentarioRepository.save(comentario));
     }
 
     // Método para eliminar comentario por su id
     @Override
-    public void eliminarComentarioById(Long idComentario) {
+    public void eliminarComentarioById(Long idComentario, Long idUsuario) {
+
+        ResponseEntity<UsuarioDTO> response = usuarioClient.obtenerUsuario(idUsuario);
+
+        if (response.getBody() == null) {
+            throw new UsuarioNotFoundException("El usuario con ID " + idUsuario + " no existe.");
+        }
+
         if (!comentarioRepository.existsById(idComentario)) {
             throw new ComentarioNotFoundException("El comentario con ID " + idComentario + " no existe.");
         }
-        comentarioRepository.deleteById(idComentario);
+        UsuarioDTO usuarioDTO = response.getBody();
+
+        if (UserRole.ADMIN.equals(usuarioDTO.getRole()) || UserRole.MODERATOR.equals(usuarioDTO.getRole())) {
+            comentarioRepository.deleteById(idComentario);
+        } else {
+            ComentarioEntity comentarioEntity = comentarioRepository.findByIdComentarioAndIdUsuario(idComentario, idUsuario)
+                    .orElseThrow(() -> new ComentarioNotFoundException(String.format("El comentario con id %s " +
+                            "no pertenece al id usuario %s",idComentario,idUsuario)));
+            comentarioRepository.delete(comentarioEntity);
+        }
     }
 
 
     // Método para eliminar todos los comentarios de una publicación
     @Override
     public void eliminarComentariosPorPublicacion(Long idPublicacion) {
-        List<ComentarioEntity> comentarios = comentarioRepository.findByIdPublicacion(idPublicacion);
+
+        if (idPublicacion <= 0) {
+            throw new IllegalNumberException("El ID debe ser positivo y no nulo");
+        }
+
+        List<ComentarioEntity> comentarios = comentarioRepository.findAllByIdPublicacion(idPublicacion).orElseGet(List::of);
 
         if (!comentarios.isEmpty()) {
             comentarioRepository.deleteAll(comentarios);
         }
+    }
+
+    @Override
+    public void eliminarComentarioPorPublicacionYUsuario(Long idPublicacion, UsuarioDTO usuarioDTO) {
+        if (idPublicacion <= 0) {
+            throw new IllegalNumberException("El ID debe ser positivo y no nulo");
+        }
+
+        List<ComentarioEntity> comentarios;
+
+        if (UserRole.ADMIN.equals(usuarioDTO.getRole()) || UserRole.MODERATOR.equals(usuarioDTO.getRole())) {
+            comentarios = comentarioRepository.findAllByIdPublicacion(idPublicacion).orElseGet(List::of);
+        } else {
+            comentarios = comentarioRepository.findAllByIdUsuario(usuarioDTO.getId()).orElseGet(List::of);
+        }
+
+        if (comentarios.isEmpty()) {
+            log.info("No hay comentarios para la publicación {}", idPublicacion);
+            return;
+        }
+        comentarioRepository.deleteAll(comentarios);
     }
 }
